Refactor "PageCrawler" (see details). - Move "scienceDirect"-relevant-code from "PageCrawler", into the new "HandleScienceDirect"-class. - Move "innerLink-checking"-code from "PageCrawler", into "UrlUtils". - Move metaDocUrl-code, JavaScriptDocLink-code & linkDebugging-code, in their own methods, inside "PageCrawler".
