+++ src/main/java/eu/openaire/doc_urls_retriever/crawler/PageCrawler.java
@@ -33,8 +33,7 @@ public class PageCrawler
 	
 	// Sciencedirect regexes. Use "find()" with those (they work best).
 	public static final Pattern META_DOC_URL = Pattern.compile("(?:<meta[\\s]*name=\"(?:citation_pdf_url|eprints.document_url)\"[\\s]*content=\")([http][\\w\\/\\.\\,\\-\\_\\%\\&\\;\\:\\~\\?\\=]+)(?:\"[\\s]*/>)");
-	public static final Pattern SCIENCEDIRECT_FINAL_DOC_URL = Pattern.compile("(?:window.location[\\s]+\\=[\\s]+\\')(.*)(?:\\'\\;)");
-	
+
 	public static final Pattern JAVASCRIPT_DOC_LINK = Pattern.compile("(?:javascript\\:pdflink.*\\')(http.+)(?:\\'\\,.*)");
 	
 	public static int totalPagesReachedCrawling = 0;	// This counts the pages which reached the crawlingStage, i.e: were not discarded in any case and waited to have their innerLinks checked.
@@ -44,59 +43,6 @@ public class PageCrawler
 	
 	public static final int timesToGiveNoInnerLinksBeforeBlocked = 5;
 	public static final int timesToGiveNoDocUrlsBeforeBlocked = 10;
-	
-	
-	public static HashSet<String> getOutgoingUrls(String pageHtml) throws JavaScriptDocLinkFoundException
-	{
-		HashSet<String> urls = new HashSet<>();
-		
-		// Get the innerLinks using "Jsoup".
-		Document document = Jsoup.parse(pageHtml);
-		Elements linksOnPage = document.select("a[href]");
-		
-		for ( Element el : linksOnPage ) {
-			String innerLink = el.attr("href");
-			if ( !innerLink.isEmpty()
-					&& !innerLink.equals("\\/") && !innerLink.equals("#")
-					&& !innerLink.startsWith("mailto:") && !innerLink.startsWith("tel:") && !innerLink.startsWith("{openurl}") ) {
-				
-				//logger.debug("InnerLink: " + innerLink);
-				String lowerCaseLink = innerLink.toLowerCase();
-				if ( lowerCaseLink.startsWith("javascript:") ) {
-					String pdfLink = null;
-					Matcher pdfLinkMatcher = JAVASCRIPT_DOC_LINK.matcher(lowerCaseLink);
-					if ( pdfLinkMatcher.matches() ) {
-						pdfLink = pdfLinkMatcher.group(1);
-						throw new JavaScriptDocLinkFoundException(pdfLink);    // If it's 'null', we treat it when handling this exception.
-					}
-					else	// It's a javaScriptLink which we don't treat.
-						continue;
-				}
-				urls.add(innerLink);
-			}
-		}
-		return urls;
-	}
-	
-	
-	public static boolean shouldNotAcceptInnerLink(String linkStr, String lowerCaseLink)
-	{
-		String lowerCaseUrl = null;
-		
-		if ( lowerCaseLink == null )
-			lowerCaseUrl = linkStr.toLowerCase();
-		else
-			lowerCaseUrl = lowerCaseLink;
-		
-		return	UrlUtils.URL_DIRECTORY_FILTER.matcher(lowerCaseUrl).matches() || UrlUtils.INNER_LINKS_KEYWORDS_FILTER.matcher(lowerCaseUrl).matches()
-				|| UrlUtils.SPECIFIC_DOMAIN_FILTER.matcher(lowerCaseUrl).matches() || UrlUtils.PLAIN_DOMAIN_FILTER.matcher(lowerCaseUrl).matches()
-				|| UrlUtils.INNER_LINKS_FILE_EXTENSION_FILTER.matcher(lowerCaseUrl).matches() || UrlUtils.INNER_LINKS_FILE_FORMAT_FILTER.matcher(lowerCaseUrl).matches()
-				|| UrlUtils.PLAIN_PAGE_EXTENSION_FILTER.matcher(lowerCaseUrl).matches()
-				|| UrlUtils.CURRENTLY_UNSUPPORTED_DOC_EXTENSION_FILTER.matcher(lowerCaseUrl).matches();	// TODO - To be removed when these docExtensions get supported.
-		
-		// The following checks are obsolete here, as we already use it inside "visit()" method. Still keep it here, as it makes our intentions clearer.
-		// !lowerCaseUrl.contains(referringPageDomain)	// Don't check this link if it belongs in a different domain than the referringPage's one.
-	}
 	
 	
 	public static void visit(String urlId, String sourceUrl, String pageUrl, HttpURLConnection conn)
@@ -111,7 +57,7 @@ public class PageCrawler
 		}
 		
 		if ( currentPageDomain.equals("linkinghub.elsevier.com") || currentPageDomain.equals("sciencedirect.com") ) {	// Be-careful if we move-on changing the retrieving of the domain of a url.
-			if ( !handleScienceDirectFamilyUrls(urlId, sourceUrl, pageUrl, currentPageDomain, conn) ) {
+			if ( !HandleScienceDirect.handleScienceDirectFamilyUrls(urlId, sourceUrl, pageUrl, currentPageDomain, conn) ) {
 				logger.warn("Problem when handling \"sciencedirect.com\" urls.");
 				UrlUtils.logQuadruple(urlId, sourceUrl, null, null, "Discarded in 'PageCrawler.visit()' method, when a 'sciencedirect.com'-url was not able to be handled correctly.", null);
 			}
@@ -131,24 +77,8 @@ public class PageCrawler
 		}
 		
 		// Check if the docLink is provided in a metaTag and connect to it directly.
-		try {
-			Matcher metaDocUrlMatcher = META_DOC_URL.matcher(pageHtml);
-			if ( metaDocUrlMatcher.find() ) {
-				String metaDocUrl = metaDocUrlMatcher.group(1);
-				if ( metaDocUrl.isEmpty() ) {
-					logger.error("Could not retrieve the metaDocUrl, continue by crawling the pageUrl.");
-				}
-				else {	// Connect to it directly.
-					//logger.debug("MetaDocUrl: " + metaDocUrl);	// DEBUG!
-					if ( !HttpConnUtils.connectAndCheckMimeType(urlId, sourceUrl, pageUrl, metaDocUrl, currentPageDomain, false, true) )	// We log the docUrl inside this method.
-						UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as the retrieved metaDocUrl was not a docUrl.", null);
-					return;
-				}
-			}
-		} catch (Exception e) {	// After connecting to the metaDocUrl.
-			UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as there was a problem with the metaTag url.", null);
-			return;
-		}
+		if ( handleMetaDocUrl(urlId, sourceUrl, pageUrl, currentPageDomain, pageHtml) )
+			return;	// The sourceUrl is already logged inside the called method.
 		
 	    // Check if we want to use AND if so, if we should run, the MLA.
 		if ( MachineLearning.useMLA ) {
@@ -161,20 +91,8 @@ public class PageCrawler
 		try {
 			currentPageLinks = getOutgoingUrls(pageHtml);
 		} catch (JavaScriptDocLinkFoundException jsdlfe) {
-			String javaScriptDocLink = jsdlfe.getMessage();
-			if ( javaScriptDocLink == null ) {
-				logger.debug("JavaScriptLink was not retrieved!");
-				UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as there was a problem retrieving its innerLinks. Its contentType is: '" + pageContentType + "'", null);
-				return;
-			}
-			try {
-				if ( !HttpConnUtils.connectAndCheckMimeType(urlId, sourceUrl, pageUrl, javaScriptDocLink, currentPageDomain, false, true) )	// We log the docUrl inside this method.
-					UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as the retrieved JavaScriptDocLink: <" + javaScriptDocLink + "> was not a docUrl.", null);
-				return;
-			} catch (Exception e) {
-				UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as the retrieved JavaScriptDocLink: <" + javaScriptDocLink + "> had connectivity problems.", null);
-				return;
-			}
+			handleJavaScriptDocLink(urlId, sourceUrl, pageUrl, currentPageDomain, pageContentType, jsdlfe);
+			return;	// This JavaScriptDOcLink is the only docLink we will ever gonna get from this page. The sourceUrl is logged inside the called method.
 		} catch (Exception e) {
 			logger.debug("Could not retrieve the innerLinks for pageUrl: " + pageUrl);
 			UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as there was a problem retrieving its innerLinks. Its contentType is: '" + pageContentType + "'", null);
@@ -191,10 +109,8 @@ public class PageCrawler
 			return;
 		}
 		
-		//Check innerLinks for debugging:
-		//if ( pageUrl.contains("<url>") )
-		/*for ( String url : currentPageLinks )
-			logger.debug(url);*/
+		//if ( pageUrl.contains(<keyWord> | <url>) )	// In case we want to print only on specific-pageTypes.
+			//printInnerLinksForDebugging(currentPageLinks);
 		
 		HashSet<String> remainingLinks = new HashSet<>();
 		String urlToCheck = null;
@@ -230,7 +146,7 @@ public class PageCrawler
             lowerCaseLink = urlToCheck.toLowerCase();
             if ( UrlUtils.DOC_URL_FILTER.matcher(lowerCaseLink).matches() )
 			{
-				if ( shouldNotAcceptInnerLink(urlToCheck, lowerCaseLink) ) {    // Avoid false-positives, such as images (a common one: ".../pdf.png").
+				if ( UrlUtils.shouldNotAcceptInnerLink(urlToCheck, lowerCaseLink) ) {    // Avoid false-positives, such as images (a common one: ".../pdf.png").
 					UrlUtils.duplicateUrls.add(urlToCheck);
 					continue;	// Disclaimer: This way we might lose some docUrls like this: "http://repositorio.ipen.br:8080/xmlui/themes/Mirage/images/Portaria-387.pdf".
 				}
@@ -268,7 +184,7 @@ public class PageCrawler
 		for ( String currentLink : remainingLinks )	// Here we don't re-check already-checked links, as they were removed.
 		{
 			// We re-check here, as, in the fast-loop not all of the links are checked against this.
-			if ( shouldNotAcceptInnerLink(currentLink, null) ) {	// If this link matches certain blackListed criteria, move on..
+			if ( UrlUtils.shouldNotAcceptInnerLink(currentLink, null) ) {	// If this link matches certain blackListed criteria, move on..
 				//logger.debug("Avoided link: " + currentLink );
 				UrlUtils.duplicateUrls.add(currentLink);
 				continue;
@@ -305,118 +221,102 @@ public class PageCrawler
 	}
 	
 	
-	/**
-	 * This method handles the JavaScriptSites of "sciencedirect.com"-family. It retrieves the docLinks hiding inside.
-	 * It returns true if the docUrl was found, otherwise, it returns false.
-	 * Note that these docUrl d not last long, since they are produced based on timestamp and jsessionid. After a while they just redirect to the pageUrl.
-	 * @param urlId
-	 * @param sourceUrl
-	 * @param pageUrl
-	 * @param pageDomain
-	 * @param conn
-	 * @return true/false
-	 */
-	public static boolean handleScienceDirectFamilyUrls(String urlId, String sourceUrl, String pageUrl, String pageDomain, HttpURLConnection conn)
+	public static HashSet<String> getOutgoingUrls(String pageHtml) throws JavaScriptDocLinkFoundException
 	{
-		try {
-			// Handle "linkinghub.elsevier.com" urls which contain javaScriptRedirect..
-			if ( pageDomain.equals("linkinghub.elsevier.com") ) {
-				//UrlUtils.elsevierLinks ++;
-				if ( (pageUrl = silentRedirectElsevierToScienseRedirect(pageUrl)) != null )
-					conn = HttpConnUtils.handleConnection(urlId, sourceUrl, pageUrl, pageUrl, pageDomain, true, false);
-				else
-					return false;
-			}
-			
-			// We now have the "sciencedirect.com" url (either from the beginning or after silentRedirect).
-			
-			logger.debug("ScienceDirect-url: " + pageUrl);
-			String html = ConnSupportUtils.getHtmlString(conn);
-			Matcher metaDocUrlMatcher = META_DOC_URL.matcher(html);
-			if ( metaDocUrlMatcher.find() )
-			{
-				String metaDocUrl = metaDocUrlMatcher.group(1);
-				if ( metaDocUrl.isEmpty() ) {
-					logger.error("Could not retrieve the metaDocUrl from a \"sciencedirect.com\" url!");
-					return false;
-				}
-				//logger.debug("MetaDocUrl: " + metaDocUrl);	// DEBUG!
+		HashSet<String> urls = new HashSet<>();
+		
+		// Get the innerLinks using "Jsoup".
+		Document document = Jsoup.parse(pageHtml);
+		Elements linksOnPage = document.select("a[href]");
+		
+		for ( Element el : linksOnPage ) {
+			String innerLink = el.attr("href");
+			if ( !innerLink.isEmpty()
+					&& !innerLink.equals("\\/") && !innerLink.equals("#")
+					&& !innerLink.startsWith("mailto:") && !innerLink.startsWith("tel:") && !innerLink.startsWith("{openurl}") ) {
 				
-				// Get the new html..
-				// We don't disconnect the previous one, since they both are in the same domain (see JavaDocs).
-				conn = HttpConnUtils.handleConnection(urlId, sourceUrl, pageUrl, metaDocUrl, pageDomain, true, false);
-				
-				//logger.debug("Url after connecting: " + conn.getURL().toString());
-				//logger.debug("MimeType: " + conn.getContentType());
-				
-				html = ConnSupportUtils.getHtmlString(conn);    // Take the new html.
-				Matcher finalDocUrlMatcher = SCIENCEDIRECT_FINAL_DOC_URL.matcher(html);
-				if ( finalDocUrlMatcher.find() )
-				{
-					String finalDocUrl = finalDocUrlMatcher.group(1);
-					if ( finalDocUrl.isEmpty() ) {
-						logger.error("Could not retrieve the finalDocUrl from a \"sciencedirect.com\" url!");
-						return false;
+				//logger.debug("InnerLink: " + innerLink);
+				String lowerCaseLink = innerLink.toLowerCase();
+				if ( lowerCaseLink.startsWith("javascript:") ) {
+					String pdfLink = null;
+					Matcher pdfLinkMatcher = JAVASCRIPT_DOC_LINK.matcher(lowerCaseLink);
+					if ( pdfLinkMatcher.matches() ) {
+						pdfLink = pdfLinkMatcher.group(1);
+						throw new JavaScriptDocLinkFoundException(pdfLink);    // If it's 'null', we treat it when handling this exception.
 					}
-					//logger.debug("FinalDocUrl: " + finalDocUrl);	// DEBUG!
-					
-					// Check and/or download the docUrl. These urls are one-time-links, meaning that after a while they will just redirect to their pageUrl.
-					if ( HttpConnUtils.connectAndCheckMimeType(urlId, sourceUrl, pageUrl, finalDocUrl, pageDomain, false, true) )    // We log the docUrl inside this method.
-						return true;
-					else {
-						logger.warn("LookedUp finalDocUrl: \"" + finalDocUrl + "\" was not an actual docUrl!");
-						return false;
-					}
-				} else {
-					logger.warn("The finalDocLink could not be found!");
-					//logger.debug("HTML-code:\n" + html);	// DEBUG!
-					return false;
+					else	// It's a javaScriptLink which we don't treat.
+						continue;
 				}
-			} else {
-				logger.warn("The metaDocLink could not be found!");	// It's possible if the document only available after paying (https://www.sciencedirect.com/science/article/pii/S1094202598900527)
-				//logger.debug("HTML-code:\n" + html);	// DEBUG!
-				return false;
+				urls.add(innerLink);
 			}
-		} catch (Exception e) {
-			logger.error("" + e);
-			return false;
 		}
-		finally {
-			// If the initial pageDomain was different from "sciencedirect.com", close the "sciencedirect.com"-connection here.
-			// Otherwise, if it came as a "sciencedirect.com", it will be closed where it was first created, meaning in "HttpConnUtils.connectAndCheckMimeType()".
-			if ( !pageDomain.equals("sciencedirect.com") )
-				conn.disconnect();	// Disconnect from the final-"sciencedirect.com"-connection.
-		}
+		return urls;
 	}
 	
 	
 	/**
-	 * This method recieves a url from "linkinghub.elsevier.com" and returns it's matched url in "sciencedirect.com".
-	 * We do this because the "linkinghub.elsevier.com" urls have a javaScript redirect inside which we are not able to handle without doing html scraping.
-	 * If there is any error this method returns the URL it first recieved.
-	 * @param linkingElsevierUrl
+	 * This method takes in the "pageHtml" of an already-connected url and checks if there is a metaDocUrl inside.
+	 * If such url exist, then it connects to it and checks if it's really a docUrl and it may also download the full-text-document, if wanted.
+	 * It returns "true" when the metaDocUrl was found and handled (independently of how it was handled),
+	 * otherwise, if the metaDocUrl was not-found, it returns "false".
+	 * @param urlId
+	 * @param sourceUrl
+	 * @param pageUrl
+	 * @param currentPageDomain
+	 * @param pageHtml
 	 * @return
 	 */
-	public static String silentRedirectElsevierToScienseRedirect(String linkingElsevierUrl)
+	public static boolean handleMetaDocUrl(String urlId, String sourceUrl, String pageUrl, String currentPageDomain, String pageHtml)
 	{
-		if ( !linkingElsevierUrl.contains("linkinghub.elsevier.com") ) // If this method was called for the wrong url, then just return it.
-			return linkingElsevierUrl;
-		
-		String idStr = null;
-		Matcher matcher = UrlUtils.URL_TRIPLE.matcher(linkingElsevierUrl);
-		if ( matcher.matches() ) {
-			idStr = matcher.group(3);
-			if ( idStr == null || idStr.isEmpty() ) {
-				logger.warn("Unexpected id-missing case for: " + linkingElsevierUrl);
-				return linkingElsevierUrl;
+		// Check if the docLink is provided in a metaTag and connect to it directly.
+		try {
+			Matcher metaDocUrlMatcher = META_DOC_URL.matcher(pageHtml);
+			if ( metaDocUrlMatcher.find() ) {
+				String metaDocUrl = metaDocUrlMatcher.group(1);
+				if ( metaDocUrl.isEmpty() ) {
+					logger.error("Could not retrieve the metaDocUrl, continue by crawling the pageUrl.");
+					return false;	// It was not handled.
+				}
+				else {	// Connect to it directly.
+					//logger.debug("MetaDocUrl: " + metaDocUrl);	// DEBUG!
+					if ( !HttpConnUtils.connectAndCheckMimeType(urlId, sourceUrl, pageUrl, metaDocUrl, currentPageDomain, false, true) )	// We log the docUrl inside this method.
+						UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as the retrieved metaDocUrl was not a docUrl.", null);
+					return true; 	// It should be the docUrl and it was handled.. so we don't continue checking the innerLink even if this wasn't a docUrl.
+				}
 			}
+			else
+				return false;	// It was not handled.
+		} catch (Exception e) {	// After connecting to the metaDocUrl.
+			UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as there was a problem with the metaTag url.", null);
+			return true;	// It was found and handled. Even if an exception was thrown, we don't want to check any other innerLinks in that page.
+		}
+	}
+	
+	
+	public static void handleJavaScriptDocLink(String urlId, String sourceUrl, String pageUrl, String currentPageDomain, String pageContentType, JavaScriptDocLinkFoundException jsdlfe)
+	{
+		String javaScriptDocLink = jsdlfe.getMessage();
+		if ( javaScriptDocLink == null ) {
+			logger.debug("JavaScriptLink was not retrieved!");
+			UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as there was a problem retrieving its innerLinks. Its contentType is: '" + pageContentType + "'", null);
 		}
 		else {
-			logger.warn("Unexpected \"URL_TRIPLE\" mismatch for: " + linkingElsevierUrl);
-			return linkingElsevierUrl;
+			try {
+				if ( !HttpConnUtils.connectAndCheckMimeType(urlId, sourceUrl, pageUrl, javaScriptDocLink, currentPageDomain, false, true) )	// We log the docUrl inside this method.
+					UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as the retrieved JavaScriptDocLink: <" + javaScriptDocLink + "> was not a docUrl.", null);
+			} catch (Exception e) {
+				UrlUtils.logQuadruple(urlId, sourceUrl, null, "unreachable", "Discarded in 'PageCrawler.visit()' method, as the retrieved JavaScriptDocLink: <" + javaScriptDocLink + "> had connectivity problems.", null);
+			}
 		}
-		
-		return ("https://www.sciencedirect.com/science/article/pii/" + idStr);
+	}
+	
+	
+	public static void printInnerLinksForDebugging(HashSet<String> currentPageLinks)
+	{
+		for ( String url : currentPageLinks ) {
+			//if ( url.contains(<keyWord> | <url>) )	// In case we want to print only specific-linkTypes.
+				logger.debug(url);
+		}
 	}
 	
 }
